+++ src/commands/index.ts
@@ -2,4 +2,5 @@ export * from './nns-help';
 export * from './nns-ping';
 export * from './nns-bug';
 export * from './nns-status';
+export * from './nns-assist';