+++ src/commands/nns-status.ts
@@ -11,57 +11,57 @@ export class StatusCommand {
     constructor(jiraApiHandler?: Jira) {
         this.jiraApiHandler = jiraApiHandler || new Jira();
     }
-    async execute(message, args, prefix = message.client.prefix, commands = message.client.commands) {
-        const helpEmbed = new MessageEmbed();
+    async execute(message, args, prefix = message.client.prefix) {
+        const statusEmbed = new MessageEmbed();
         let jiraResponse;
-        helpEmbed.type = 'rich';
+        statusEmbed.type = 'rich';
         if (args && args.length > 0) {
 
             try {
                 jiraResponse = await this.jiraApiHandler.getTicketStatus(args[0]);
 
                 if (jiraResponse && jiraResponse.hasOwnProperty('errorMessages')) {
-                    helpEmbed.setColor('#FF0000')
+                    statusEmbed.setColor('#FF0000')
                         .setTitle(`Invalid Ticket or Permission is denied!`)
                         .setDescription(`${jiraResponse['errorMessages'][0]}`)
                         .setTimestamp();
                 }
                 if (jiraResponse && jiraResponse.hasOwnProperty('fields')) {
-                    helpEmbed.setColor('#00ff00')
+                    statusEmbed.setColor('#00ff00')
                         .setTitle(`${jiraResponse['fields'].project.name}`)
                         .setURL(`${jiraResponse['fields'].project.self}`)
                         .setDescription(`${jiraResponse['fields'].summary}`)
                         .setTimestamp();
-                    helpEmbed.addField('Assignee', jiraResponse['fields'].assignee
+                    statusEmbed.addField('Assignee', jiraResponse['fields'].assignee
                         ? `${jiraResponse['fields'].assignee.displayName}`
                         : 'Not Assigned', true);
                     if (jiraResponse['fields'].reporter
                         && jiraResponse['fields'].reporter.hasOwnProperty('displayName')) {
-                        helpEmbed.addField('Reporter', `${jiraResponse['fields'].reporter.displayName}`, true);
+                        statusEmbed.addField('Reporter', `${jiraResponse['fields'].reporter.displayName}`, true);
                     }
-                    helpEmbed.addField('Status', `${jiraResponse['fields'].status.name}`, false);
+                    statusEmbed.addField('Status', `${jiraResponse['fields'].status.name}`, false);
                 }
                 try {
-                    message.channel.send(helpEmbed);
+                    message.channel.send(statusEmbed);
                 }
                 catch (e) {
                     console.error(e);
                 }
-                return helpEmbed;
+                return statusEmbed;
             } catch (e) {
                 console.error(e);
             }
         }
-        helpEmbed.setColor('#F8AA2A')
+        statusEmbed.setColor('#F8AA2A')
             .setTitle('JIRA Ticket Status')
             .setDescription(this.description)
             .addField(`${prefix}${this.alias}`, this.man, true)
             .setTimestamp();
         try {
-            message.channel.send(helpEmbed);
+            message.channel.send(statusEmbed);
         } catch (e) {
             console.error(e);
         }
-        return helpEmbed;
+        return statusEmbed;
     }
 }